Extract load_symbol_from_dylib and dlsym_proc_macros into a new module
